backend/requirements.txt
# requirements.txt
# HealthMitra - AI-Driven Public Health Chatbot
# Core Python dependencies

# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# AI/NLP Services
openai==1.3.7
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
rasa==3.6.17
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz

# Translation
googletrans==4.0.0-rc1
translate==3.6.1
langdetect==1.0.9

# Text-to-Speech & Voice
gtts==2.5.1
pyttsx3==2.90
vosk==0.3.45
pyaudio==0.2.14
SpeechRecognition==3.10.0

# WhatsApp & SMS Integration
whatsapp-cloud-api==0.2.2
twilio==8.10.0
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Caching & Performance
redis==5.0.1
aioredis==2.0.1
cachetools==5.3.2

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7
python-dotenv==1.0.0

# Data Processing & Utilities
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
python-dateutil==2.8.2
pytz==2023.3

# File Handling
Pillow==10.1.0
python-magic==0.4.27

# Monitoring & Logging
sentry-sdk==1.35.0
structlog==23.2.0
prometheus-client==0.18.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
responses==0.24.1
freezegun==1.2.2

# Code Quality & Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
autoflake==2.2.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings[python]==0.23.0

# Other Utilities
click==8.1.7
tqdm==4.66.1
colorama==0.4.6
jinja2==3.1.2
markdown==3.5.1
pyyaml==6.0.1